build: clap v4~~> v3 -- v4 does not support color output
Exploring using clap for now, but intend to switch to a front-end in
another language later.  For some scripts, ironically, using Python may
be worth it as (a) there is nominally a good crate that eases FFI'ing
into Python (b) Python has an *excellent* CLI framework (`Typer`) and (c) the
`pip3 install <wheel>` process is familiar and easy for users in mind and can be done
without creating a private server.  Caution-Point: adding a binary
dependency to a python-wheel may complicate its distribution and may
require mutliple, binary-build specific, wheels <-- tbd
(also need to look into maintenance costs -- especially if rust library is
not published on public repo like pypi)